[hack] fix rgba32 textures
due to how wigwam was developed, rgba32 textures to some capacity were only passing half their bytes (2,048 vs 4,096). This fixs those, albit in a way it shouldn't be (backpack doesn't do this), but for now, i will assume this is a proper fix, as a real fix appears to be a bad rewrite, which i really don't want to deal with. Model dumps also are fixed with this.